{"schema":"decodex/commit/1","summary":"Split MCP HTTP message modules","authority":"XY-1115"}
