+++ apps/decodex/src/mcp/http/message.rs
@@ -1,265 +1,9 @@
-use std::str;
+mod request;
+mod response;
+mod rpc;
 
-use serde_json::{self, Value};
-
-use crate::prelude::Result;
-
-use super::{
-	super::{MCP_PROTOCOL_VERSION, MCP_SESSION_HEADER},
-	MCP_CORS_ALLOW_HEADERS, MCP_CORS_ALLOW_METHODS,
+pub(in crate::mcp) use self::{
+	request::{McpHttpRequest, http_content_length, http_header_end},
+	response::{McpHttpResponse, mcp_http_response_for_server_responses},
+	rpc::{initialize_response_succeeded, json_rpc_method_name},
 };
-
-pub(super) struct McpHttpRequest {
-	pub(super) method: String,
-	pub(super) path: String,
-	pub(super) headers: Vec<(String, String)>,
-	pub(super) body: Vec<u8>,
-}
-
-impl McpHttpRequest {
-	pub(super) fn parse(request: &[u8]) -> std::result::Result<Self, McpHttpResponse> {
-		let Some(header_end) = http_header_end(request) else {
-			return Err(McpHttpResponse::empty("400 Bad Request"));
-		};
-		let header_text = str::from_utf8(&request[..header_end])
-			.map_err(|_| McpHttpResponse::empty("400 Bad Request"))?;
-		let mut lines = header_text.split("\r\n");
-		let request_line = lines.next().ok_or_else(|| McpHttpResponse::empty("400 Bad Request"))?;
-		let mut request_parts = request_line.split_whitespace();
-		let method = request_parts
-			.next()
-			.ok_or_else(|| McpHttpResponse::empty("400 Bad Request"))?
-			.to_owned();
-		let path = request_parts
-			.next()
-			.ok_or_else(|| McpHttpResponse::empty("400 Bad Request"))?
-			.to_owned();
-		let version =
-			request_parts.next().ok_or_else(|| McpHttpResponse::empty("400 Bad Request"))?;
-
-		if !version.starts_with("HTTP/1.") {
-			return Err(McpHttpResponse::empty("505 HTTP Version Not Supported"));
-		}
-
-		let mut headers = Vec::new();
-
-		for line in lines {
-			if line.is_empty() {
-				continue;
-			}
-
-			let Some((name, value)) = line.split_once(':') else {
-				return Err(McpHttpResponse::empty("400 Bad Request"));
-			};
-
-			headers.push((name.trim().to_owned(), value.trim().to_owned()));
-		}
-
-		let content_length = http_content_length(&request[..header_end])
-			.map_err(|_| McpHttpResponse::empty("400 Bad Request"))?;
-		let body_start = header_end + 4;
-		let body_end = body_start.saturating_add(content_length);
-
-		if request.len() < body_end {
-			return Err(McpHttpResponse::empty("400 Bad Request"));
-		}
-
-		Ok(Self { method, path, headers, body: request[body_start..body_end].to_vec() })
-	}
-
-	pub(super) fn header(&self, name: &str) -> Option<&str> {
-		self.headers
-			.iter()
-			.find(|(header, _)| header.eq_ignore_ascii_case(name))
-			.map(|(_, value)| value.as_str())
-	}
-
-	pub(super) fn accepts_sse(&self) -> bool {
-		header_contains(self.header("Accept"), "text/event-stream")
-	}
-
-	pub(super) fn content_type_is_json(&self) -> bool {
-		header_contains(self.header("Content-Type"), "application/json")
-	}
-}
-
-pub(super) struct McpHttpResponse {
-	pub(super) status: &'static str,
-	pub(super) content_type: Option<&'static str>,
-	pub(super) headers: Vec<(&'static str, String)>,
-	pub(super) body: Vec<u8>,
-}
-
-impl McpHttpResponse {
-	pub(super) fn empty(status: &'static str) -> Self {
-		Self { status, content_type: None, headers: Vec::new(), body: Vec::new() }
-	}
-
-	pub(super) fn empty_with_session(status: &'static str, session_id: Option<String>) -> Self {
-		let mut response = Self::empty(status);
-
-		response.add_session_header(session_id);
-
-		response
-	}
-
-	pub(super) fn json(value: Value, session_id: Option<String>) -> Result<Self> {
-		let body = serde_json::to_vec(&value)?;
-		let mut response = Self {
-			status: "200 OK",
-			content_type: Some("application/json"),
-			headers: vec![("Cache-Control", String::from("no-store"))],
-			body,
-		};
-
-		response.add_session_header(session_id);
-
-		Ok(response)
-	}
-
-	pub(super) fn json_error(status: &'static str, value: Value) -> Self {
-		let body =
-			serde_json::to_vec(&value)
-				.unwrap_or_else(|_| {
-					br#"{"jsonrpc":"2.0","id":null,"error":{"code":-32603,"message":"Internal error"}}"#.to_vec()
-				});
-
-		Self {
-			status,
-			content_type: Some("application/json"),
-			headers: vec![("Cache-Control", String::from("no-store"))],
-			body,
-		}
-	}
-
-	pub(super) fn sse(responses: Vec<Value>, session_id: Option<String>) -> Result<Self> {
-		let mut body = Vec::new();
-
-		for response in responses {
-			let line = serde_json::to_string(&response)?;
-
-			body.extend_from_slice(b"event: message\n");
-			body.extend_from_slice(b"data: ");
-			body.extend_from_slice(line.as_bytes());
-			body.extend_from_slice(b"\n\n");
-		}
-
-		let mut response = Self {
-			status: "200 OK",
-			content_type: Some("text/event-stream"),
-			headers: vec![
-				("Cache-Control", String::from("no-store")),
-				("X-Accel-Buffering", String::from("no")),
-			],
-			body,
-		};
-
-		response.add_session_header(session_id);
-
-		Ok(response)
-	}
-
-	fn add_session_header(&mut self, session_id: Option<String>) {
-		if let Some(session_id) = session_id {
-			self.headers.push((MCP_SESSION_HEADER, session_id));
-		}
-	}
-
-	pub(super) fn add_cors_headers(&mut self, origin: Option<String>) {
-		let Some(origin) = origin else {
-			return;
-		};
-
-		self.headers.push(("Access-Control-Allow-Origin", origin));
-		self.headers.push(("Vary", String::from("Origin")));
-		self.headers.push(("Access-Control-Allow-Methods", String::from(MCP_CORS_ALLOW_METHODS)));
-		self.headers.push(("Access-Control-Allow-Headers", String::from(MCP_CORS_ALLOW_HEADERS)));
-		self.headers.push(("Access-Control-Expose-Headers", String::from(MCP_SESSION_HEADER)));
-	}
-
-	pub(super) fn into_bytes(self) -> Result<Vec<u8>> {
-		let mut response = Vec::new();
-
-		response.extend_from_slice(format!("HTTP/1.1 {}\r\n", self.status).as_bytes());
-		response.extend_from_slice(b"Connection: close\r\n");
-		response.extend_from_slice(format!("Content-Length: {}\r\n", self.body.len()).as_bytes());
-
-		if let Some(content_type) = self.content_type {
-			response.extend_from_slice(format!("Content-Type: {content_type}\r\n").as_bytes());
-		}
-
-		for (name, value) in self.headers {
-			response.extend_from_slice(format!("{name}: {value}\r\n").as_bytes());
-		}
-
-		response.extend_from_slice(b"\r\n");
-		response.extend_from_slice(&self.body);
-
-		Ok(response)
-	}
-}
-
-pub(super) fn mcp_http_response_for_server_responses(
-	responses: Vec<Value>,
-	wants_sse: bool,
-	session_id: Option<String>,
-) -> Result<McpHttpResponse> {
-	if responses.is_empty() {
-		return Ok(McpHttpResponse::empty_with_session("202 Accepted", session_id));
-	}
-	if wants_sse {
-		return McpHttpResponse::sse(responses, session_id);
-	}
-
-	McpHttpResponse::json(
-		responses.into_iter().next().unwrap_or_else(|| serde_json::json!({})),
-		session_id,
-	)
-}
-
-pub(in crate::mcp) fn http_header_end(bytes: &[u8]) -> Option<usize> {
-	bytes.windows(4).position(|window| window == b"\r\n\r\n")
-}
-
-pub(super) fn http_content_length(header_bytes: &[u8]) -> Result<usize> {
-	let header_text = str::from_utf8(header_bytes)?;
-
-	for line in header_text.split("\r\n").skip(1) {
-		let Some((name, value)) = line.split_once(':') else {
-			continue;
-		};
-
-		if name.trim().eq_ignore_ascii_case("Content-Length") {
-			return Ok(value.trim().parse()?);
-		}
-	}
-
-	Ok(0)
-}
-
-fn header_contains(header: Option<&str>, value: &str) -> bool {
-	header
-		.map(|header| {
-			header.split(',').any(|item| {
-				item.trim().split(';').next().is_some_and(|item| item.eq_ignore_ascii_case(value))
-			})
-		})
-		.unwrap_or(false)
-}
-
-pub(super) fn json_rpc_method_name(body: &str) -> Option<String> {
-	serde_json::from_str::<Value>(body)
-		.ok()
-		.and_then(|value| value.get("method").and_then(Value::as_str).map(str::to_owned))
-}
-
-pub(super) fn initialize_response_succeeded(responses: &[Value]) -> bool {
-	responses.iter().any(|response| {
-		response.get("error").is_none()
-			&& response
-				.get("result")
-				.and_then(|result| result.get("protocolVersion"))
-				.and_then(Value::as_str)
-				== Some(MCP_PROTOCOL_VERSION)
-	})
-}
